Okay, testing on this version has finished.
I am going to pull the files out of here and work on them separately
for the second build. Some things I tried implementing are not
compatible with this version and the code is very messy and not a lot
has been done. I am going to restart this with a blank template