Implimented the popup screen for all HeaderButtons, with page specific content support.
Moved the logic for displaying the Modal popup screen into the HeaderButton object definition. The header class uses the icon name passed in for the HeaderButton to determine which content to display on the corresponding Modal popup screen. It now works for every header button. Repurposed the NewEntryDialogue object as a container for page specific content on the the add new entry page. Next objective for me will be adding in redux support in several places. And after that, implimenting our API in pantry.