Cleaned up view of popup screens & toggled off the TabNavigator animation. (#16)
* Cleaned up view of popup screens

Left aligned the back button and changed its icon color to white. Changed title on new entry popup screen. Turned off animation for TabNavigator. We can see if this helps the performance or not, but I was worried that the animation was slowing things down. I left the swipeing between screens enabled.

* Moved title into header bar. Formatted its styling.

Added a title prop to the HeaderButton popup screens. The title is displayed withing the blue header bar, to the right of the back button, with adequate spacing between them.