Added Search bar support to Pantry screens
Adding in a SearchBar Component from 'react-native-elements'. This replaces the search button that was previosuly on the view all screen, and adds to a more intuitive user experience. I have styled and customized the SearchBar containers to mesh well with the application. No backend functionality in place for this feature yet.