Updated View All Pantry Toggle Button
Changed component to the Button class in 'react-native-elements' as opposed to the one in 'react-native'. Much better in my opinion. Also updated styling and color of the button. Color is Amber 500. We mentioned the use of a blue, yellow, and gray scheme. This seems like a good place for an accent color.