Correcting rounded corners
The images in dashboard were no longer responding to the border radius due to the fact that they had children. I have fixed this by using 'position: absolute' to stack the transparent background and text headlines on top of the images instead of having them be children of the image. Corresponding style changes have been made to support this refactor.